main screen, meal functionality, timers service

+++ app/src/main/java/by/slavintodron/babyhelper/entity/weatherEntity/WeatherData.kt
@@ -5,7 +5,7 @@ data class WeatherData(
     val dt: Int,
     val dt_txt: String,
     val main: Main,
-    val pop: Int,
+    val pop: Float,
     val sys: Sys,
     val visibility: Int,
     val weather: List<Weather>,
